requirements.txt
aiogram==3.11.0
aiosqlite==0.20.0
load-dotenv==0.1.0
SQLAlchemy==2.0.40